requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==24.2.0
bcrypt==4.2.0
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.3.2
CTkToolTip==0.8
customtkinter==5.2.2
darkdetect==0.8.0
distro==1.9.0
exceptiongroup==1.2.2
Faker==28.1.0
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
jiter==0.5.0
multidict==6.0.5
openai==1.43.0
outcome==1.3.0.post0
packaging==24.1
pillow==10.4.0
playwright==1.46.0
psycopg2==2.9.9
pydantic==2.8.2
pydantic_core==2.20.1
pyee==11.1.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
selenium==4.24.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
style==1.1.0
tkinter-tooltip==3.1.0
tqdm==4.66.5
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
update==0.0.1
urllib3==2.2.2
websocket-client==1.8.0
wsproto==1.2.0
yarl==1.9.7
